solidigm: Clean up compatibility fixes in solidigm plugin (#104)
Convert to using %"PRIu64" format string for platform-agnostic 64-bit value handling.

+++ plugins/solidigm/solidigm-smart.c
@@ -163,12 +163,12 @@ static void smart_log_item_print(struct nvme_additional_smart_log_item *item,
 			le32_to_cpu(item->thermal_throttle.count));
 		return;
 	case 0xF3:
-		printf("gain0: %u, loss0: %u, gain1: %u, loss1: %u, legacy:%llu\n",
+		printf("gain0: %u, loss0: %u, gain1: %u, loss1: %u, legacy:%"PRIu64"\n",
 			le16_to_cpu(pll_item->gainCount0),
 			le16_to_cpu(pll_item->lossCount0),
 			le16_to_cpu(pll_item->gainCount1),
 			le16_to_cpu(pll_item->lossCount1),
-			(unsigned long long)int48_to_long(item->raw));
+			int48_to_long(item->raw));
 		return;
 	default:
 		printf("%"PRIu64"\n", int48_to_long(item->raw));
